pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "confee"
version = "0.0.dev"  # Auto-updated by GitHub Actions when tag is pushed
description = "Configuration Enhanced & Easy ☕️ - A Hydra-like configuration parser helper package"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "JunSeok Kim", email = "infend@gmail.com"}
]
keywords = ["configuration", "config", "parser", "hydra", "pydantic", "yaml"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "devtools>=0.12.2",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
cli = ["typer>=0.9.0"]
remote = ["aiohttp>=3.8.0"]
toml = []
all = ["typer>=0.9.0", "aiohttp>=3.8.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/bestend/confee"
Documentation = "https://github.com/bestend/confee#readme"
Repository = "https://github.com/bestend/confee.git"
Issues = "https://github.com/bestend/confee/issues"

[tool.setuptools]
packages = ["confee"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
confee = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py312', 'py313', 'py314']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "D", "SIM"]
ignore = [
    "D100", "D101", "D102", "D104", "D105", "D107", "D200", "D205", "D415", "D417",
    "N818", "N806", "N814",
    "E501",  # Line too long (conflicts with black)
    "B904", "B007",  # Exception handling and loop control
    "E721",  # Use isinstance instead of type comparison
    "SIM117",  # Use a single with statement
    "UP006", "UP035",  # Use new-style type hints
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = false
no_error_summary = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=confee --cov-report=term-missing"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
